317. glxinfo needs libGLU (Marc La France).  316. Rage128 fix to Cards database (Marc La France).  315. Minor fix to PCI resource overlap handling (Marc La France).  314. Loader code simplification and IA-64 cache flushes (Marc La France).  313. Workaround in the ATI driver for troublesome interaction between loader       and compiler optimisation (Marc La France).  312. Warning fixes for `gcc -fno-builtin`, which appears to be the default on       some systems (Marc La France).
+ more of #301, tags, warnings and build fixes.

+++ lib/GL/mesa/src/drv/r128/r128_ioctl.h
@@ -1,4 +1,4 @@
-/* $XFree86: xc/lib/GL/mesa/src/drv/r128/r128_ioctl.h,v 1.1 2000/12/04 19:21:46 dawes Exp $ */
+/* $XFree86: xc/lib/GL/mesa/src/drv/r128/r128_ioctl.h,v 1.2 2001/01/08 01:07:21 martin Exp $ */
 /**************************************************************************
 
 Copyright 1999, 2000 ATI Technologies Inc. and Precision Insight, Inc.,
@@ -141,7 +141,7 @@ do {									\
 #define ALIGN_NEXT_ELT( rmesa )						\
 do {									\
    rmesa->next_elt = (GLushort *)					\
-      (((GLuint)rmesa->next_elt + 7) & ~0x7);				\
+      (((unsigned long)rmesa->next_elt + 7) & ~0x7);			\
    rmesa->next_elt = (GLushort *)					\
       ((GLubyte *)rmesa->next_elt + R128_INDEX_PRIM_OFFSET);		\
 } while (0)
